Fix test frozen time
As we are adding new channels as demo data, these channels will have a process_end_date
defined to today when the module is installed, and as tests were run with a time frozen
in 2023, there were always channels where the process end time was after the scheduled
date of the picking.

Using relative frozen time should help in keeping the tests successful through different
times.